+++ processor/src/decoder/mod.rs
@@ -327,11 +327,12 @@ impl Process {
         let mem_addr = self.stack.get(0);
         // The callee hash is stored in memory, and the address is specified on the top of the
         // stack.
-        let callee_hash = self
-            .chiplets
-            .memory
-            .read_word(self.system.ctx(), mem_addr, self.system.clk(), error_ctx)
-            .map_err(ExecutionError::MemoryError)?;
+        let callee_hash = self.chiplets.memory.read_word(
+            self.system.ctx(),
+            mem_addr,
+            self.system.clk(),
+            error_ctx,
+        )?;
 
         let (addr, hashed_block) = self.chiplets.hasher.hash_control_block(
             EMPTY_WORD,
@@ -365,11 +366,12 @@ impl Process {
         let mem_addr = self.stack.get(0);
         // The callee hash is stored in memory, and the address is specified on the top of the
         // stack.
-        let callee_hash = self
-            .chiplets
-            .memory
-            .read_word(self.system.ctx(), mem_addr, self.system.clk(), error_ctx)
-            .map_err(ExecutionError::MemoryError)?;
+        let callee_hash = self.chiplets.memory.read_word(
+            self.system.ctx(),
+            mem_addr,
+            self.system.clk(),
+            error_ctx,
+        )?;
 
         // Note: other functions end in "executing a Noop", which
         // 1. ensures trace capacity,