+++ processor/src/fast/memory.rs
@@ -3,7 +3,7 @@ use alloc::{collections::BTreeMap, vec::Vec};
 use miden_air::RowIndex;
 use vm_core::{EMPTY_WORD, Felt, WORD_SIZE, Word, ZERO};
 
-use crate::{ContextId, ExecutionError, MemoryAddress, MemoryError, errors::ErrorContext};
+use crate::{ContextId, ErrorContext, MemoryAddress, MemoryError};
 
 /// The memory for the processor.
 ///
@@ -24,7 +24,7 @@ impl Memory {
     ///
     /// # Errors
     /// - Returns an error if the provided address is out-of-bounds.
-    pub fn read_element(&mut self, ctx: ContextId, addr: Felt) -> Result<Felt, ExecutionError> {
+    pub fn read_element(&mut self, ctx: ContextId, addr: Felt) -> Result<Felt, MemoryError> {
         let element = self.read_element_impl(ctx, clean_addr(addr)?).unwrap_or(ZERO);
         Ok(element)
     }
@@ -38,9 +38,9 @@ impl Memory {
         ctx: ContextId,
         addr: Felt,
         clk: RowIndex,
-    ) -> Result<&Word, ExecutionError> {
+    ) -> Result<Word, MemoryError> {
         let addr = clean_addr(addr)?;
-        let word = self.read_word_impl(ctx, addr, Some(clk))?.unwrap_or(&EMPTY_WORD);
+        let word = self.read_word_impl(ctx, addr, Some(clk))?.unwrap_or(EMPTY_WORD);
 
         Ok(word)
     }
@@ -54,7 +54,7 @@ impl Memory {
         ctx: ContextId,
         addr: Felt,
         element: Felt,
-    ) -> Result<(), ExecutionError> {
+    ) -> Result<(), MemoryError> {
         let (word_addr, idx) = split_addr(clean_addr(addr)?);
 
         self.memory
@@ -84,7 +84,7 @@ impl Memory {
         addr: Felt,
         clk: RowIndex,
         word: Word,
-    ) -> Result<(), ExecutionError> {
+    ) -> Result<(), MemoryError> {
         let addr = enforce_word_aligned_addr(ctx, clean_addr(addr)?, Some(clk))?;
         self.memory.insert((ctx, addr), word);
 
@@ -135,9 +135,9 @@ impl Memory {
         ctx: ContextId,
         addr: u32,
         clk: Option<RowIndex>,
-    ) -> Result<Option<&Word>, ExecutionError> {
+    ) -> Result<Option<Word>, MemoryError> {
         let addr = enforce_word_aligned_addr(ctx, addr, clk)?;
-        let word = self.memory.get(&(ctx, addr));
+        let word = self.memory.get(&(ctx, addr)).copied();
 
         Ok(word)
     }
@@ -150,14 +150,10 @@ impl Memory {
 ///
 /// # Errors
 /// - Returns an error if the provided address is out-of-bounds.
-fn clean_addr(addr: Felt) -> Result<u32, ExecutionError> {
+fn clean_addr(addr: Felt) -> Result<u32, MemoryError> {
     let addr = addr.as_int();
-    addr.try_into().map_err(|_| {
-        ExecutionError::MemoryError(MemoryError::address_out_of_bounds(
-            addr,
-            &ErrorContext::default(),
-        ))
-    })
+    addr.try_into()
+        .map_err(|_| MemoryError::address_out_of_bounds(addr, &ErrorContext::default()))
 }
 
 /// Splits the provided address into the word address and the index within the word.
@@ -180,19 +176,16 @@ fn enforce_word_aligned_addr(
     ctx: ContextId,
     addr: u32,
     clk: Option<RowIndex>,
-) -> Result<u32, ExecutionError> {
+) -> Result<u32, MemoryError> {
     if addr % WORD_SIZE as u32 != 0 {
         return match clk {
-            Some(clk) => Err(ExecutionError::MemoryError(MemoryError::unaligned_word_access(
+            Some(clk) => Err(MemoryError::unaligned_word_access(
                 addr,
                 ctx,
                 Felt::from(clk.as_u32()),
                 &ErrorContext::default(),
-            ))),
-            None => Err(ExecutionError::MemoryError(MemoryError::UnalignedWordAccessNoClk {
-                addr,
-                ctx,
-            })),
+            )),
+            None => Err(MemoryError::UnalignedWordAccessNoClk { addr, ctx }),
         };
     }